backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
pillow
python-multipart
opencv-python
python-dotenv
